+++ src/main/java/io/hops/metadata/hdfs/dal/LargeOnDiskInodeDataAccess.java
@@ -7,7 +7,7 @@ import io.hops.metadata.common.EntityDataAccess;
  * Created by salman on 3/10/16.
  */
 public interface LargeOnDiskInodeDataAccess<T> extends DBFileDataAccess<T>, EntityDataAccess {
-  T get(int var1, int size) throws StorageException;
+  T get(long var1, int size) throws StorageException;
   int countUniqueFiles() throws StorageException;
 }
