+++ src/main/java/io/hops/metadata/hdfs/entity/MetadataLogEntry.java
@@ -16,10 +16,10 @@
 package io.hops.metadata.hdfs.entity;
 
 public class MetadataLogEntry {
-  private final int datasetId;
-  private final int inodeId;
-  private final int inodePartitionId;
-  private final int inodeParentId;
+  private final long datasetId;
+  private final long inodeId;
+  private final long inodePartitionId;
+  private final long inodeParentId;
   private final String inodeName;
   private int logicalTime;
   private final Operation operation;
@@ -32,8 +32,8 @@ public class MetadataLogEntry {
     CHANGEDATASET
   }
 
-  public MetadataLogEntry(int datasetId, int inodeId, int inodePartitionId,
-      int inodeParentId, String inodeName, int logicalTime, Operation
+  public MetadataLogEntry(long datasetId, long inodeId, long inodePartitionId,
+      long inodeParentId, String inodeName, int logicalTime, Operation
       operation) {
     this.datasetId = datasetId;
     this.inodeId = inodeId;
@@ -44,11 +44,11 @@ public class MetadataLogEntry {
     this.operation = operation;
   }
 
-  public int getDatasetId() {
+  public long getDatasetId() {
     return datasetId;
   }
 
-  public int getInodeId() {
+  public long getInodeId() {
     return inodeId;
   }
 
@@ -64,11 +64,11 @@ public class MetadataLogEntry {
     return (short) operation.ordinal();
   }
 
-  public int getInodePartitionId() {
+  public long getInodePartitionId() {
     return inodePartitionId;
   }
 
-  public int getInodeParentId() {
+  public long getInodeParentId() {
     return inodeParentId;
   }
